requirements.txt
alembic
click          
Flask         
Flask-Assets   
Flask-Babel
Flask-Script
httplib2     
itsdangerous   
Jinja2       
Mako        
MarkupSafe   
oauth2
pycparser 
python-editor
requests
webassets
Werkzeug   
gunicorn
beautifulsoup4
line-bot-sdk